:tropical_drink: Token Sessão Acções na págia Eventos
* not/going
* add/remove commentarios

+++ Source Code/PHP/addremoveComment.php
@@ -1,9 +1,9 @@
 <?php
 
-if (isset($_POST['NewComment'])) {
+if (isset($_POST['NewComment']) && $_SESSION['csrf'] == $_POST['csrf']) {
 	newComment($_POST['Event_ID'],$_POST['User_ID'],$_POST['NewCommentText']);	
 }
-if (isset($_POST['DeleteComment'])) {
+if (isset($_POST['DeleteComment']) && $_SESSION['csrf'] == $_POST['csrf']) {
 	if(checkIfCommentExists($_POST['Comment_ID']))
 		deleteComment($_POST['Comment_ID']);
